Feat: create functions for connection, insertion and find to mongo
